refactor(core): unify sse for Planar and globe
Fix wfs example, different subdivision after change computing preSSE.
