+++ examples/wfs.html
@@ -160,7 +160,7 @@
                 projection: 'EPSG:4326',
                 ipr: 'IGN',
                 format: 'application/json',
-                zoom: { min: 5, max: 5 },
+                zoom: { min: 4, max: 4 },
                 extent: {
                     west: 4.568,
                     east: 5.18,
